requirements.txt
streamlit
deepface
opencv-python-headless
tensorflow
keras
numpy
pillow
matplotlib
